# development/backend/openapi.json
{
  "components": {
    "schemas": {
      "AccountBriefingGroup": {
        "properties": {
          "connected_account_id": {
            "format": "uuid",
            "title": "Connected Account Id",
            "type": "string"
          },
          "gmail_address": {
            "title": "Gmail Address",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BriefingCard"
            },
            "title": "Items",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "syncing": {
            "title": "Syncing",
            "type": "boolean"
          }
        },
        "required": [
          "connected_account_id",
          "gmail_address",
          "status",
          "syncing",
          "items"
        ],
        "title": "AccountBriefingGroup",
        "type": "object"
      },
...
      "CreateLabelRequest": {
        "properties": {
          "connected_account_id": {
            "format": "uuid",
            "title": "Connected Account Id",
            "type": "string"
          },
          "hidden": {
            "default": false,
            "title": "Hidden",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "order_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Index"
          }
        },
        "required": [
          "connected_account_id",
          "name"
        ],
        "title": "CreateLabelRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ItemStateResult": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "remind_later_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remind Later At"
          },
          "seen": {
            "title": "Seen",
            "type": "boolean"
          },
          "seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seen At"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "message_id",
          "seen",
          "seen_at",
          "remind_later_at",
          "version",
          "updated_at"
        ],
        "title": "ItemStateResult",
        "type": "object"
      },
...
      "MessageDetail": {
        "description": "GET /messages/{id} readonly view.\n\nmutation action field(mark_read/archive/label)\uac00 \uc5c6\uace0 AI reason field\ub3c4 \uae30\ubcf8\uc73c\ub85c \uc5c6\ub2e4.\nbriefing.md \"Read API\" \u00a7negative / \u00a7\ube48\uc0c1\ud0dc \uae30\uc900\uc774\ub2e4.",
        "properties": {
          "connected_account_id": {
            "format": "uuid",
            "title": "Connected Account Id",
            "type": "string"
          },
          "done": {
            "title": "Done",
            "type": "boolean"
          },
          "excerpt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt Text"
          },
          "gmail_message_id": {
            "title": "Gmail Message Id",
            "type": "string"
          },
          "gmail_thread_id": {
            "title": "Gmail Thread Id",
            "type": "string"
          },
          "gmail_url": {
            "title": "Gmail Url",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "importance_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance Band"
          },
          "received_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "summary_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Text"
          }
        },
        "required": [
          "id",
          "connected_account_id",
          "gmail_message_id",
          "gmail_thread_id",
          "gmail_url",
          "subject",
          "sender",
          "received_at",
          "excerpt_text",
          "summary_text",
          "importance_band",
          "done"
        ],
        "title": "MessageDetail",
        "type": "object"
      },
      "MoveMessageRequest": {
        "properties": {
          "label_id": {
            "format": "uuid",
            "title": "Label Id",
            "type": "string"
          }
        },
        "required": [
          "label_id"
        ],
        "title": "MoveMessageRequest",
        "type": "object"
      },
      "MoveMessageResult": {
        "properties": {
          "correction_signal_id": {
            "format": "uuid",
            "title": "Correction Signal Id",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "service_label_id": {
            "format": "uuid",
            "title": "Service Label Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "correction_signal_id",
          "message_id",
          "service_label_id",
          "version"
        ],
        "title": "MoveMessageResult",
        "type": "object"
      },
...
      "RemindLaterRequest": {
        "properties": {
          "remind_at": {
            "format": "date-time",
            "title": "Remind At",
            "type": "string"
          }
        },
        "required": [
          "remind_at"
        ],
        "title": "RemindLaterRequest",
        "type": "object"
      },
      "ReminderResult": {
        "properties": {
          "briefing_item_state_id": {
            "format": "uuid",
            "title": "Briefing Item State Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reactivated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reactivated At"
          },
          "remind_at": {
            "format": "date-time",
            "title": "Remind At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "briefing_item_state_id",
          "remind_at",
          "reactivated_at",
          "status"
        ],
        "title": "ReminderResult",
        "type": "object"
      },
...
      "ServiceLabel": {
        "properties": {
          "connected_account_id": {
            "format": "uuid",
            "title": "Connected Account Id",
            "type": "string"
          },
          "gmail_label_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gmail Label Id"
          },
          "gmail_label_name": {
            "title": "Gmail Label Name",
            "type": "string"
          },
          "hidden": {
            "title": "Hidden",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "order_index": {
            "title": "Order Index",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "name",
          "order_index",
          "hidden",
          "updated_at",
          "connected_account_id",
          "gmail_label_id",
          "gmail_label_name"
        ],
        "title": "ServiceLabel",
        "type": "object"
      },
...
      "UpcomingStorage": {
        "properties": {
          "this_week": {
            "items": {
              "$ref": "#/components/schemas/UpcomingReminderEntry"
            },
            "title": "This Week",
            "type": "array"
          },
          "today": {
            "items": {
              "$ref": "#/components/schemas/UpcomingReminderEntry"
            },
            "title": "Today",
            "type": "array"
          },
          "tomorrow": {
            "items": {
              "$ref": "#/components/schemas/UpcomingReminderEntry"
            },
            "title": "Tomorrow",
            "type": "array"
          }
        },
        "required": [
          "today",
          "tomorrow",
          "this_week"
        ],
        "title": "UpcomingStorage",
        "type": "object"
      },
      "UpdateLabelRequest": {
        "properties": {
          "hidden": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hidden"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "order_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Index"
          }
        },
        "title": "UpdateLabelRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Maily API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/briefing/items/{briefing_item_id}/remind_later": {
      "post": {
        "operationId": "remind_later",
        "parameters": [
          {
            "in": "path",
            "name": "briefing_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Briefing Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemindLaterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Remind Later",
        "tags": [
          "briefing"
        ]
      }
    },
    "/briefing/items/{briefing_item_id}/seen": {
      "post": {
        "operationId": "mark_item_seen",
        "parameters": [
          {
            "in": "path",
            "name": "briefing_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Briefing Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemStateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Mark Item Seen",
        "tags": [
          "briefing"
        ]
      }
    },
    "/briefing/today": {
      "get": {
        "operationId": "today_briefing",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountBriefingGroup"
                  },
                  "title": "Response Today Briefing",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Today Briefing",
        "tags": [
          "briefing"
        ]
      }
    },
...
    "/labels": {
      "get": {
        "operationId": "get_labels",
        "parameters": [
          {
            "in": "query",
            "name": "include_hidden",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Hidden",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceLabel"
                  },
                  "title": "Response Get Labels",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Get Labels",
        "tags": [
          "labels"
        ]
      },
      "post": {
        "operationId": "create_label",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceLabel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Create Label",
        "tags": [
          "labels"
        ]
      }
    },
    "/labels/{label_id}": {
      "patch": {
        "operationId": "patch_label",
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Label Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceLabel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Patch Label",
        "tags": [
          "labels"
        ]
      }
    },
    "/messages/{message_id}": {
      "get": {
        "operationId": "message_detail",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Message Detail",
        "tags": [
          "briefing"
        ]
      }
    },
    "/messages/{message_id}/move": {
      "post": {
        "operationId": "move_message",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveMessageResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Move Message",
        "tags": [
          "labels"
        ]
      }
    },
...
    "/storage/upcoming": {
      "get": {
        "operationId": "storage_upcoming",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpcomingStorage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Maily error envelope"
          }
        },
        "summary": "Storage Upcoming",
        "tags": [
          "briefing"
        ]
      }
    }
  }
}